Slim up the preamble to rely on ctypes capability
This patch begins an attempt to slim up the amount of "tweaks" that must
be used in the preamble to make things work smoothly.  Rather than using
a special POINTER function and a special c_void class, this patch changes
void* returns to be converted into real ctypes.c_void_p instances.  These
instances make casting easier (as was one of the intents of the special
c_void class) but also do this without introducing a new POINTER class.

By avoiding the non-ctypes POINTER function (and what it was doing),
ctypes' auto-byref calling works again.  Hence, the experience of using
ctypesgen creates less of a layer over ctypes that requires a different
documentation in order to  use.

Fixes davidjamesca/ctypesgen#26