Keep pointer qualifiers around so to decide whether to return c_char_p or String
ctypes already converts items returned as c_char_p as native python
strings.  If a function returns a (const char*), there is no reason to want
to return an instance of the MutableString class.  This is part of the
effort to eliminate as much of the "tweaks" in preamble as possible in
order to make the usage of ctypesgen generated code be closer to documented
ctypes.