Add option to run lexer in optimized mode
Previously, the lexer always attempted to run in optimized mode.  This is a
particular problem if ctypesgen is installed as a system-level package
where write access to the lextab.py file is limited.  This change does not
select the optimized mode by default, thus not requiring root-access to
execute.

It may be necessary to use a highly diverse h-file as input to ensure that
the lexer has seen all types of tokens so that it does not have to be
re-written.