Adds a few minor tests for main entry point
This patch tests ctypesgen.main.main() as used in the local script 'run.py'
to ensure that simple commandline processing works as expected.  The actual
script installed is similar to the local script, but is slightly more
general and is allowed to be placed outside of the code tree.  That script
is autogenerated by setuptools.

This resolves #68, where at least some testing of the commandline interface
was requested.