Use sys.executable to run separate script in testsuite
In 5171ac75c00b, new tests were added to testsuite to do at least some
exercise of the main entry point to ctypesgen, as it is installed.  This
patch improves 5171ac75c00b by using sys.executable as the executable to
run Python script with.  It is critical to use sys.executable, since it is
problematic to guess what the commandline executable is for a particular
running instance of Python.

This patch improves commits 7006cb27a8c0 and 5171ac75c00b.

Fixes davidjamesca/ctypesgen#79
Closes davidjamesca/ctypesgen#80