Allow __GNUC__ if desired by the user
Ctypesgen implicitly undefines '__GNUC__' while invoking the C preprocessor.
This interferes with parsing well-written code that contains gcc-specific
code within "#ifdef __GNUC__".

A common example is the "packed" attribute that ctypesgen supports parsing.
Currently all instances of "packed" defined within "#ifdef __GNUC__" are being
stripped-off when ctypesgen invokes the C preprocesser with -U __GNUC__.

This commit introduces a command-line parameter "--allow-gnu-c" that allows
the user of ctypesgen to override this legacy behaviour of ctypesgen.

This ensures that the default behaviour of ctypesgen remains unchanged.
Users of ctypesgen who require "__GNUC__" to continue to be defined when
ctypesgen parses their code, can invoke it as "ctypesgen --allow-gnu-c ..."

NOTE: It is not clear why specifying -D __GNUC__ does NOT override the implicit
-U __GNUC__ i.e. it still results in __GNUC__ NOT being defined. This is likely
a result of the order in which the define/undefine are passed/processed.
Instead of toying with the order in the parser code to achieve the desired
behaviour, this change introduces the new parameter "--allow-gnu-c" to enable
the user of ctypesgen to explicitly and reliably control the desired behaviour.